Bug 1013662 - Fix bad implicit conversion constructors in MFBT; r=froydnj,Waldo
UltraBlame original commit: d7dc39d162348d21bd2b51b3a3b56fa2fe969dd5

+++ mfbt/MathAlgorithms.h
@@ -25,8 +25,8 @@ EuclidGCD(IntegerType a, IntegerType b)
 {
   
   
-  MOZ_ASSERT(a > 0);
-  MOZ_ASSERT(b > 0);
+  MOZ_ASSERT(a > IntegerType(0));
+  MOZ_ASSERT(b > IntegerType(0));
 
   while (a != b) {
     if (a > b) {
